# docs/swagger.yaml
definitions:
...
  db.Camera:
    properties:
      cam_id:
        type: integer
      cam_ip:
        type: string
      cam_name:
        type: string
      cam_password:
        type: string
      cam_port:
        type: integer
      cam_type:
        type: string
      cam_user:
        type: string
      direction:
        type: string
      extra:
        type: object
      is_enabled:
        type: boolean
      zone_in_id:
        type: integer
      zone_out_id:
        type: integer
    required:
    - cam_id
    - cam_ip
    - cam_name
    - cam_password
    - cam_port
    - cam_type
    - cam_user
    - direction
    - zone_in_id
    - zone_out_id
    type: object
...
info:
  contact: {}
  title: Find Your Car
  version: 0.0.9
paths:
...
  /fyc/cameras:
    delete:
      description: Delete a camera by setting the is_deleted flag to true
      parameters:
      - description: Camera ID
        in: query
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Camera deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid camera ID
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to delete camera
          schema:
            additionalProperties: true
            type: object
      summary: Delete a camera
      tags:
      - Cameras
    get:
      parameters:
      - description: 'State of cameras: enabled or deleted'
        in: query
        name: state
        type: string
      - description: Include extra information if 'yes'
        in: query
        name: extra
        type: string
      responses:
        "200":
          description: List of cameras or a single camera
          schema:
            $ref: '#/definitions/db.Camera'
        "400":
          description: 'Bad request: Invalid camera ID'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: No cameras found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get camera data
      tags:
      - Cameras
    post:
      consumes:
      - application/json
      description: Add a new camera to the database
      parameters:
      - description: Camera data
        in: body
        name: Camera
        required: true
        schema:
          $ref: '#/definitions/db.Camera'
      produces:
      - application/json
      responses:
        "201":
          description: Camera created successfully
          schema:
            $ref: '#/definitions/db.Camera'
        "400":
          description: Invalid request payload
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to create a new camera
          schema:
            additionalProperties: true
            type: object
      summary: Add a new camera
      tags:
      - Cameras
    put:
      consumes:
      - application/json
      description: Update an existing camera by ID
      parameters:
      - description: Camera ID
        in: query
        name: id
        required: true
        type: integer
      - description: Updated camera data
        in: body
        name: Camera
        required: true
        schema:
          $ref: '#/definitions/db.Camera'
      produces:
      - application/json
      responses:
        "200":
          description: Camera updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request payload or ID mismatch
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Camera not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to update camera
          schema:
            additionalProperties: true
            type: object
      summary: Update a camera by ID
      tags:
      - Cameras
...
securityDefinitions:
  BearerAuth3rdParty:
    description: 'Authorization token for third-party section (Ensure the token is
      in this format: Bearer token)'
    in: header
    name: Authorization
    type: apiKey
  BearerAuthBackOffice:
    description: 'Authorization token for back-office section (Ensure the token is
      in this format: Bearer token)'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"